timpa

+++ libraries/BluetoothSerial/examples/DiscoverConnect/DiscoverConnect.ino
@@ -15,7 +15,6 @@
  * use ESP_SPP_SEC_NONE or ESP_SPP_SEC_ENCRYPT|ESP_SPP_SEC_AUTHENTICATE in connect() if remote side has Authentication: False
  */
 
-#include <Arduino.h>
 #include <map>
 #include <BluetoothSerial.h>
 
@@ -64,7 +63,7 @@ void setup() {
         BTAdvertisedDevice *device = btDeviceList->getDevice(i);
         Serial.printf(" ----- %s  %s %d\n", device->getAddress().toString().c_str(), device->getName().c_str(), device->getRSSI());
         std::map<int, std::string> channels = SerialBT.getChannels(device->getAddress());
-        Serial.printf("scanned for services, found %lu\n", (unsigned long)channels.size());
+        Serial.printf("scanned for services, found %zu\n", channels.size());
         for (auto const &entry : channels) {
           Serial.printf("     channel %d (%s)\n", entry.first, entry.second.c_str());
         }
