+++ libraries/Update/examples/OTAWebUpdater/OTAWebUpdater.ino
@@ -1,4 +1,3 @@
-#include <Arduino.h>
 #include <WiFi.h>
 #include <WebServer.h>
 #include <ESPmDNS.h>
@@ -84,7 +83,7 @@ void handleUpdate() {
       return;
     }
 
-    Serial.printf("Receiving Update: %s, Size: %lu\n", upload.filename.c_str(), (unsigned long)fsize);
+    Serial.printf("Receiving Update: %s, Size: %zu\n", upload.filename.c_str(), fsize);
     if (!Update.begin(fsize)) {
       otaDone = 0;
       Update.printError(Serial);
@@ -97,7 +96,7 @@ void handleUpdate() {
     }
   } else if (authenticated && upload.status == UPLOAD_FILE_END) {
     if (Update.end(true)) {
-      Serial.printf("Update Success: %lu bytes\nRebooting...\n", (unsigned long)upload.totalSize);
+      Serial.printf("Update Success: %zu bytes\nRebooting...\n", upload.totalSize);
     } else {
       Serial.printf("%s\n", Update.errorString());
       otaDone = 0;
@@ -132,7 +131,7 @@ void webServerInit() {
 
 void everySecond() {
   if (otaDone > 1) {
-    Serial.printf("ota: %u%%\n", otaDone);
+    Serial.printf("ota: %d%%\n", otaDone);
   }
 }
