+++ libraries/WebServer/examples/WebUpdate/WebUpdate.ino
@@ -2,7 +2,6 @@
   To upload through terminal you can use: curl -F "image=@firmware.bin" esp8266-webupdate.local/update
 */
 
-#include <Arduino.h>
 #include <WiFi.h>
 #include <NetworkClient.h>
 #include <WebServer.h>
@@ -83,7 +82,7 @@ void setup(void) {
           }
         } else if (authenticated && upload.status == UPLOAD_FILE_END) {
           if (Update.end(true)) {  //true to set the size to the current progress
-            Serial.printf("Update Success: %lu\nRebooting...\n", (unsigned long)upload.totalSize);
+            Serial.printf("Update Success: %zu\nRebooting...\n", upload.totalSize);
           } else {
             Update.printError(Serial);
           }
